make solvers work on chain reference:
This way you can change a chain in between calls, to make sure that no allocations happen inadvertently an extra function updateInternalDataStructures is added to the SolverI class that needs to be implemented for each solver. This function needs to be called if the number of segments or number of joints has changed before calling a new JntToCart/CartToJnt. All solvers also need to check that the internal state is valid before calculating.

This patch also adds more library-wide solver error codes and introduces them where appropriate.

Signed-off-by: Ruben Smits <ruben.smits@intermodalics.eu>